+++ Helper_Tests/combinatorics_tests.cpp
@@ -7,7 +7,7 @@
 
 
 TEST(Combinatorics, CombinationN0) {
-	for (int n = 1; n < 10; ++n) {
+	for (int8_t n = 1; n < 10; ++n) {
 		auto combo_list = get_combinations(n, 0);
 		EXPECT_EQ(combo_list.size(), 1);
 		EXPECT_EQ(combo_list.begin()->size(), 0);
@@ -15,7 +15,7 @@ TEST(Combinatorics, CombinationN0) {
 }
 
 TEST(Combinatorics, CombinationNN) {
-	for (int n = 1; n < 10; ++n) {
+	for (int8_t n = 1; n < 10; ++n) {
 		auto combo_list = get_combinations(n, n);
 		EXPECT_EQ(combo_list.size(), 1);
 		EXPECT_EQ(combo_list.begin()->size(), n);
@@ -23,7 +23,7 @@ TEST(Combinatorics, CombinationNN) {
 }
 
 TEST(Combinatorics, CombinationN1) {
-	for (int n = 1; n < 10; ++n) {
+	for (int8_t n = 1; n < 10; ++n) {
 		auto combo_list = get_combinations(n, 1);
 		EXPECT_EQ(combo_list.size(), n);
 		for (const auto& combo : combo_list) {
@@ -33,7 +33,7 @@ TEST(Combinatorics, CombinationN1) {
 }
 
 TEST(Combinatorics, CombinationN2) {
-	for (int n = 2; n < 10; ++n) {
+	for (int8_t n = 2; n < 10; ++n) {
 		auto combo_list = get_combinations(n, 2);
 		EXPECT_EQ(combo_list.size(), n*(n-1)/2);
 		for (const auto& combo : combo_list) {
@@ -43,7 +43,7 @@ TEST(Combinatorics, CombinationN2) {
 }
 
 TEST(Combinatorics, CombinationN3) {
-	for (int n = 3; n < 10; ++n) {
+	for (int8_t n = 3; n < 10; ++n) {
 		auto combo_list = get_combinations(n, 3);
 		EXPECT_EQ(combo_list.size(), n * (n - 1) * (n - 2) / 6);
 		for (const auto& combo : combo_list) {
@@ -53,7 +53,7 @@ TEST(Combinatorics, CombinationN3) {
 }
 
 TEST(Combinatorics, PermutationN0) {
-	for (int n = 1; n < 10; ++n) {
+	for (int8_t n = 1; n < 10; ++n) {
 		auto perm_list = get_permutations(n, 0);
 		// For perm(n, 0) we should get a single empty permutation.
 		EXPECT_EQ(perm_list.size(), 1);
@@ -62,11 +62,11 @@ TEST(Combinatorics, PermutationN0) {
 }
 
 TEST(Combinatorics, PermutationN1) {
-	for (int n = 1; n < 10; ++n) {
+	for (int8_t n = 1; n < 10; ++n) {
 		auto perm_list = get_permutations(n, 1);
 		// For perm(n, 1) we should get n 1-element permutations, which are each in [0, n) in order.
 		EXPECT_EQ(perm_list.size(), n);
-		for (int i = 0; i < n; ++i) {
+		for (int8_t i = 0; i < n; ++i) {
 			EXPECT_EQ(perm_list[i].size(), 1);
 			EXPECT_EQ(perm_list[i].front(), i);
 		}
@@ -74,10 +74,10 @@ TEST(Combinatorics, PermutationN1) {
 }
 
 TEST(Combinatorics, PermutationNN) {
-	for (int n = 1; n < 10; ++n) {
+	for (int8_t n = 1; n < 10; ++n) {
 		auto perm_list = get_permutations(n, n);
 		EXPECT_EQ(perm_list.size(), factorial(n));
-		for (int i = 0; i < n; ++i) {
+		for (int8_t i = 0; i < n; ++i) {
 			EXPECT_EQ(perm_list[i].size(), n);
 		}
 		// Verify each oredered permutation is unique by putting them into a set
